Move get_environment() to base class
