+++ client/verta/verta/registry/entities/_modelversion.py
@@ -33,7 +33,7 @@ from verta._internal_utils import (
 from verta import utils
 
 from verta import _blob, code, data_types
-from verta.environment import _Environment, Python
+from verta.environment import _Environment
 from verta.monitoring import profiler
 from verta.tracking.entities._entity import _MODEL_ARTIFACTS_ATTR_KEY
 from verta.tracking.entities import _deployable_entity
@@ -607,22 +607,6 @@ class RegisteredModelVersion(_deployable_entity._DeployableEntity):
         self._fetch_with_no_cache()
         self._msg.ClearField("environment")
         self._update(self._msg, method="PUT")
-
-    def get_environment(self):
-        """
-        Gets the environment of this Model Version.
-
-        Returns
-        -------
-        :class:`~verta.environment.Python`
-            Environment of this ModelVersion.
-
-        """
-        self._refresh_cache()
-        if not self.has_environment:
-            raise RuntimeError("environment was not previously set.")
-
-        return Python._from_proto(self._msg)
 
     def _get_url_for_artifact(self, key, method, artifact_type=0, part_num=0):
         if method.upper() not in ("GET", "PUT"):
