+++ client/verta/verta/tracking/entities/_experimentrun.py
@@ -36,7 +36,7 @@ from verta.dataset.entities import (
 from verta import data_types
 from verta import deployment
 from verta import utils
-from verta.environment import Python
+from verta.environment import _Environment
 
 from ._entity import _MODEL_ARTIFACTS_ATTR_KEY
 from ._deployable_entity import _DeployableEntity
@@ -1651,9 +1651,10 @@ class ExperimentRun(_DeployableEntity):
         return _utils.unravel_observations(self._msg.observations)
 
     def log_environment(self, env, overwrite=False):
-        if not isinstance(env, Python):
+        if not isinstance(env, _Environment):
             raise TypeError(
-                "`env` must be of type Python, not {}".format(type(env)))
+                "`env` must be of type Environment, not {}".format(type(env))
+            )
 
         if self.has_environment and not overwrite:
             raise ValueError(
@@ -1671,22 +1672,6 @@ class ExperimentRun(_DeployableEntity):
             self._fetch_with_no_cache()
         else:
             _utils.raise_for_http_error(response)
-
-    def get_environment(self):
-        """
-        Gets the environment of this Experiment Run.
-
-        Returns
-        -------
-        :class:`~verta.environment.Python`
-            Environment of this ExperimentRun.
-
-        """
-        self._refresh_cache()
-        if not self.has_environment:
-            raise RuntimeError("environment was not previously set")
-
-        return Python._from_proto(self._msg)
 
     def log_modules(self, paths, search_path=None):
         """
